requirements.txt
tqdm
ipykernel
atlasify
networkx
seaborn
pyyaml
click
lightning==2.1.2
torch-geometric==2.4.0
-f https://data.pyg.org/whl/torch-2.1.0+cu121.html
torch-scatter
torch-cluster
--extra-index-url https://pypi.nvidia.com
cudf-cu12 
cugraph-cu12
cuml-cu12
wandb
